+++ src/main/steps/common/task-list/components/progress-bar/config/ca_respondent.ts
@@ -1,9 +1,11 @@
 /* eslint-disable @typescript-eslint/explicit-module-boundary-types */
+import { CaseWithId } from '../../../../../../app/case/case';
 import { State } from '../../../../../../app/case/definition';
+import { ProgressBarProps } from '../../../../../../steps/common/task-list/definitions';
 import { getPartyDetails } from '../../../../../../steps/tasklistresponse/utils';
 import { isFinalOrderIssued, progressBarStage } from '../utils';
 
-export const CA_RESPONDENT = [
+export const CA_RESPONDENT: ProgressBarProps[] = [
   {
     ...progressBarStage.applicationSubmitted,
     isComplete: () => true,
@@ -12,10 +14,8 @@ export const CA_RESPONDENT = [
   {
     ...progressBarStage.responseSubmitted,
     isInProgress: (caseData, userDetails) => {
-      const respondent = getPartyDetails(caseData, userDetails.id);
-      if (respondent?.response.citizenFlags?.isResponseInitiated) {
-        return true;
-      }
+      const respondent = getPartyDetails(caseData as CaseWithId, userDetails.id);
+      return !!respondent?.response.citizenFlags?.isResponseInitiated;
     },
     isComplete: (caseData, userDetails) =>
       caseData.citizenResponseC7DocumentList?.find(doc => doc.value.createdBy === userDetails.id) !== undefined,
@@ -23,9 +23,11 @@ export const CA_RESPONDENT = [
   {
     ...progressBarStage.hearingAndCourtOrders,
     isInProgress: caseData =>
-      caseData.orderCollection ||
-      caseData.state === State.DECISION_OUTCOME ||
-      caseData.state === State.PREPARE_FOR_HEARING_CONDUCT_HEARING,
+      !!(
+        caseData.orderCollection ||
+        caseData.state === State.DECISION_OUTCOME ||
+        caseData.state === State.PREPARE_FOR_HEARING_CONDUCT_HEARING
+      ),
     isComplete: isFinalOrderIssued,
   },
   {
