+++ src/main/steps/common/task-list/components/progress-bar/config/da_applicant_respondent.ts
@@ -1,17 +1,21 @@
 /* eslint-disable @typescript-eslint/explicit-module-boundary-types */
+
 import { CaseWithId } from '../../../../../../app/case/case';
 import { SelectTypeOfOrderEnum, State } from '../../../../../../app/case/definition';
+import { ProgressBarProps } from '../../../../../../steps/common/task-list/definitions';
 import { isCaseClosed } from '../../../utils';
 import { progressBarStage } from '../utils';
 
-export const DA_APPLICANT_RESPONDENT = [
+export const DA_APPLICANT_RESPONDENT: ProgressBarProps[] = [
   { ...progressBarStage.caseOpened, isComplete: () => true },
   {
     ...progressBarStage.hearingAndCourtOrders,
     isInProgress: (userCase: Partial<CaseWithId>) =>
       userCase &&
-      (userCase.orderCollection ||
-        [State.DECISION_OUTCOME, State.PREPARE_FOR_HEARING_CONDUCT_HEARING].includes(userCase.state!)),
+      !!(
+        userCase.orderCollection ||
+        [State.DECISION_OUTCOME, State.PREPARE_FOR_HEARING_CONDUCT_HEARING].includes(userCase.state!)
+      ),
     isComplete: (userCase: Partial<CaseWithId>) => userCase.selectTypeOfOrder === SelectTypeOfOrderEnum.finl,
   },
   {
