+++ src/main/steps/common/task-list/components/tasklist/config/ca_applicant.ts
@@ -1,6 +1,8 @@
 /* eslint-disable @typescript-eslint/explicit-module-boundary-types */
 import { CaseWithId } from '../../../../../../app/case/case';
+import { Respondent } from '../../../../../../app/case/definition';
 import { UserDetails } from '../../../../../../app/controller/AppRequest';
+import { interpolate } from '../../../../../../steps/common/string-parser';
 import {
   APPLICANT_CHECK_ANSWERS,
   APPLICANT_DETAILS_KNOWN,
@@ -12,11 +14,13 @@ import {
   APPLICANT_YOURHEARINGS_HEARINGS,
   C100_DOWNLOAD_APPLICATION,
   C100_START,
+  TASKLIST_RESPONSE_TO_CA,
 } from '../../../../../../steps/urls';
+import { SectionContent, Task, TaskListConfigProps } from '../../../definitions';
 import { isCaseClosed, isCaseLinked, isDraftCase, isRepresentedBySolicotor } from '../../../utils';
 import { StateTags, TaskListSection, Tasks, getContents, hasAnyHearing, hasAnyOrder } from '../utils';
 
-export const CA_APPLICANT = [
+export const CA_APPLICANT: TaskListConfigProps[] = [
   {
     id: TaskListSection.ABOUT_YOU,
     content: getContents.bind(null, TaskListSection.ABOUT_YOU),
@@ -27,7 +31,7 @@ export const CA_APPLICANT = [
         !isRepresentedBySolicotor(caseData as CaseWithId, userDetails.id)
       );
     },
-    tasks: [
+    tasks: (): Task[] => [
       {
         id: Tasks.EDIT_YOUR_CONTACT_DETAILS,
         href: (caseData: Partial<CaseWithId>) => `${APPLICANT_CHECK_ANSWERS}/${caseData.id}`,
@@ -59,14 +63,14 @@ export const CA_APPLICANT = [
   {
     id: TaskListSection.YOUR_APPLICATION,
     content: getContents.bind(null, TaskListSection.YOUR_APPLICATION),
-    tasks: [
+    tasks: (): Task[] => [
       {
         id: Tasks.CHILD_ARRANGEMENT_APPLICATION,
         href: (caseData: Partial<CaseWithId>) => {
           if (!caseData) {
             return C100_START;
           }
-          return caseData.c100RebuildReturnUrl;
+          return caseData.c100RebuildReturnUrl!;
         },
         stateTag: (caseData: Partial<CaseWithId>) => {
           if (!caseData) {
@@ -88,7 +92,7 @@ export const CA_APPLICANT = [
     id: TaskListSection.YOUR_DOCUMENTS,
     content: getContents.bind(null, TaskListSection.YOUR_DOCUMENTS),
     show: isCaseLinked,
-    tasks: [
+    tasks: (): Task[] => [
       {
         id: Tasks.UPLOAD_DOCUMENTS,
         href: () => APPLICANT_UPLOAD_DOCUMENT_LIST_URL,
@@ -112,7 +116,7 @@ export const CA_APPLICANT = [
     id: TaskListSection.YOUR_ORDERS,
     content: getContents.bind(null, TaskListSection.YOUR_ORDERS),
     show: isCaseLinked,
-    tasks: [
+    tasks: (): Task[] => [
       {
         id: Tasks.VIEW_ORDERS,
         href: () => APPLICANT_ORDERS_FROM_THE_COURT,
@@ -127,10 +131,16 @@ export const CA_APPLICANT = [
     ],
   },
   {
+    id: TaskListSection.THE_RESPONSE,
+    content: getContents.bind(null, TaskListSection.THE_RESPONSE),
+    show: isCaseLinked,
+    tasks: (caseData, content): Task[] => generateTheResponseTasks(caseData, content),
+  },
+  {
     id: TaskListSection.YOUR_HEARING,
     content: getContents.bind(null, TaskListSection.YOUR_HEARING),
     show: isCaseLinked,
-    tasks: [
+    tasks: (): Task[] => [
       {
         id: Tasks.VIEW_HEARING_DETAILS,
         href: (caseData: Partial<CaseWithId>) => `${APPLICANT_YOURHEARINGS_HEARINGS}/${caseData.id}`,
@@ -145,3 +155,36 @@ export const CA_APPLICANT = [
     ],
   },
 ];
+
+const isResponsePresent = (caseData: Partial<CaseWithId>, respondent: Respondent) => {
+  return caseData.respondentDocsList?.find(
+    documents => documents.value.c7Document?.partyName === respondent.value.firstName + ' ' + respondent.value.lastName
+  );
+};
+
+const generateTheResponseTasks = (caseData: Partial<CaseWithId>, content: SectionContent): Task[] => {
+  const tasks: Task[] = [];
+
+  caseData.respondents?.forEach(respondent => {
+    tasks.push({
+      id: Tasks.THE_RESPONSE_PDF,
+      linkText: interpolate(content?.tasks[Tasks.THE_RESPONSE_PDF]!.linkText, {
+        respondentPosition: String(caseData.respondents!.indexOf(respondent) + 1),
+      }),
+      href: () => {
+        const respondentName = respondent.value.firstName + ' ' + respondent.value.lastName;
+        //TODO change to use url parameter when citizen document upload changes are merged
+        return `${TASKLIST_RESPONSE_TO_CA}?name=${respondentName}`;
+      },
+      stateTag: () => {
+        return isResponsePresent(caseData, respondent) ? StateTags.READY_TO_VIEW : StateTags.NOT_AVAILABLE_YET;
+      },
+      show: () => caseData && !isDraftCase(caseData),
+      disabled: () => {
+        return !isResponsePresent(caseData, respondent);
+      },
+    });
+  });
+
+  return tasks;
+};
