+++ src/main/steps/common/task-list/components/tasklist/config/da_applicant.ts
@@ -1,6 +1,7 @@
 /* eslint-disable @typescript-eslint/explicit-module-boundary-types */
 import { CaseWithId } from '../../../../../../app/case/case';
 import { UserDetails } from '../../../../../../app/controller/AppRequest';
+import { Task, TaskListConfigProps } from '../../../../../../steps/common/task-list/definitions';
 import { isCaseClosed, isRepresentedBySolicotor } from '../../../../../../steps/common/task-list/utils';
 import {
   APPLICANT_ATTENDING_THE_COURT,
@@ -26,14 +27,14 @@ import {
   hasAnyOrder,
 } from '../utils';
 
-export const DA_APPLICANT = [
+export const DA_APPLICANT: TaskListConfigProps[] = [
   {
     id: TaskListSection.ABOUT_YOU,
     content: getContents.bind(null, TaskListSection.ABOUT_YOU),
     show: (caseData: Partial<CaseWithId>, userDetails: UserDetails) => {
       return !isCaseClosed(caseData) && !isRepresentedBySolicotor(caseData as CaseWithId, userDetails.id);
     },
-    tasks: [
+    tasks: (): Task[] => [
       {
         id: Tasks.KEEP_YOUR_DETAILS_PRIVATE,
         href: (caseData: Partial<CaseWithId>) => `${APPLICANT_DETAILS_KNOWN}/${caseData.id}`,
@@ -67,7 +68,7 @@ export const DA_APPLICANT = [
   {
     id: TaskListSection.YOUR_APPLICATION,
     content: getContents.bind(null, TaskListSection.YOUR_APPLICATION),
-    tasks: [
+    tasks: (): Task[] => [
       {
         id: Tasks.YOUR_APPLICATION_PDF,
         href: () => YOUR_APPLICATION_FL401,
@@ -85,7 +86,7 @@ export const DA_APPLICANT = [
   {
     id: TaskListSection.YOUR_HEARING,
     content: getContents.bind(null, TaskListSection.YOUR_HEARING),
-    tasks: [
+    tasks: (): Task[] => [
       {
         id: Tasks.VIEW_HEARING_DETAILS,
         href: (caseData: Partial<CaseWithId>) =>
@@ -103,7 +104,7 @@ export const DA_APPLICANT = [
   {
     id: TaskListSection.YOUR_DOCUMENTS,
     content: getContents.bind(null, TaskListSection.YOUR_DOCUMENTS),
-    tasks: [
+    tasks: (): Task[] => [
       {
         id: Tasks.UPLOAD_DOCUMENTS,
         href: () => APPLICANT_UPLOAD_DOCUMENT_LIST_URL,
@@ -124,7 +125,7 @@ export const DA_APPLICANT = [
   {
     id: TaskListSection.YOUR_ORDERS,
     content: getContents.bind(null, TaskListSection.YOUR_ORDERS),
-    tasks: [
+    tasks: (): Task[] => [
       {
         id: Tasks.VIEW_ORDERS,
         href: caseData => (hasAnyOrder(caseData) ? APPLICANT_ORDERS_FROM_THE_COURT : '#'),
