+++ src/main/steps/common/task-list/components/tasklist/index.ts
@@ -1,14 +1,23 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 import _ from 'lodash';
 
 import { CaseWithId } from '../../../../../app/case/case';
 import { CaseType, PartyType } from '../../../../../app/case/definition';
 import { UserDetails } from '../../../../../app/controller/AppRequest';
+import {
+  HintConfig,
+  HyperLinkConfig,
+  PreparedTask,
+  SectionContent,
+  StateTagsConfig,
+  Task,
+  TaskListConfig,
+  TaskListConfigProps,
+} from '../../definitions';
 
 import tasklistConfig from './config/index';
 import { StateTags, getStateTagLabel } from './utils';
 
-const stateTagsConfig = {
+const stateTagsConfig: StateTagsConfig = {
   [StateTags.NOT_STARTED_YET]: {
     label: getStateTagLabel.bind(null, StateTags.NOT_STARTED_YET),
     className: 'govuk-tag--red',
@@ -48,10 +57,11 @@ const stateTagsConfig = {
   },
   [StateTags.VIEW]: {
     label: getStateTagLabel.bind(null, StateTags.VIEW),
+    className: 'govuk-tag--dark-blue',
   },
 };
 
-const taskListConfig = {
+const taskListConfig: TaskListConfig = {
   [CaseType.C100]: {
     [PartyType.APPLICANT]: tasklistConfig.CA_APPLICANT,
     [PartyType.RESPONDENT]: tasklistConfig.CA_RESPONDENT,
@@ -67,7 +77,7 @@ export const getTaskListConfig = (
   userDetails: UserDetails,
   partyType: PartyType,
   language: string
-): Record<string, any>[] => {
+): TaskListConfigProps[] => {
   let caseType = caseData?.caseTypeOfApplication;
   if (!caseType && partyType === PartyType.APPLICANT) {
     caseType = CaseType.C100;
@@ -84,14 +94,15 @@ export const getTaskListConfig = (
         return {
           id: section.id,
           heading: _content.heading,
-          tasks: section.tasks
+          tasks: section
+            .tasks(caseData, _content)
             .map(task => {
               if (!task.hasOwnProperty('show') || (task.show instanceof Function && task.show(caseData, userDetails))) {
-                const config = prepareTaskListConfig(task, caseData, userDetails, _content, language, partyType);
-
-                prepareHintConfig(task, caseData, userDetails, config, _content);
-                prepareOpenInAnotherTabConfig(task, config);
-                return config;
+                return {
+                  ...prepareTaskListConfig(task, caseData, userDetails, _content, language, partyType),
+                  ...prepareHintConfig(task, caseData, userDetails, _content),
+                  ...prepareHyperLinkConfig(task),
+                };
               }
               return null;
             })
@@ -107,14 +118,15 @@ export const getTaskListConfig = (
       return config !== null;
     });
 };
+
 const prepareTaskListConfig = (
-  task: any,
+  task: Task,
   caseData: Partial<CaseWithId>,
   userDetails: UserDetails,
-  _content: any,
+  _content: SectionContent,
   language: string,
   partyType: PartyType
-) => {
+): PreparedTask => {
   const stateTag = task.stateTag(caseData, userDetails);
   const _stateTagConfig = stateTagsConfig?.[stateTag];
 
@@ -127,7 +139,7 @@ const prepareTaskListConfig = (
 
   const config = {
     id: task.id,
-    linkText: _content?.tasks[task.id]?.linkText,
+    linkText: task?.linkText ?? _content?.tasks[task.id]?.linkText,
     href: task.href(caseData, userDetails),
     disabled: _.isFunction(task?.disabled) ? task.disabled(caseData, userDetails) : false,
     stateTag: {
@@ -139,26 +151,21 @@ const prepareTaskListConfig = (
 };
 
 const prepareHintConfig = (
-  task: any,
+  task: Task,
   caseData: Partial<CaseWithId>,
   userDetails: UserDetails,
-  config: { id: any; linkText: any; href: any; disabled: any; stateTag: { label: any; className: any } },
-  _content: any
-) => {
-  if (task?.showHint && task.showHint instanceof Function && task.showHint(caseData, userDetails)) {
-    Object.assign(config, {
-      hintText: _content?.tasks[task.id]?.hintText,
-    });
-  }
+  _content: SectionContent
+): HintConfig => {
+  return {
+    hintText:
+      task?.showHint && task.showHint instanceof Function && task.showHint(caseData, userDetails)
+        ? _content?.tasks[task.id]?.hintText
+        : null,
+  };
 };
 
-const prepareOpenInAnotherTabConfig = (
-  task: any,
-  config: { id: any; linkText: any; href: any; disabled: any; stateTag: { label: any; className: any } }
-) => {
-  if (task.openInAnotherTab) {
-    Object.assign(config, {
-      openInAnotherTab: task.openInAnotherTab,
-    });
-  }
+const prepareHyperLinkConfig = (task: Task): HyperLinkConfig => {
+  return {
+    openInAnotherTab: task.openInAnotherTab,
+  };
 };
