+++ src/main/steps/common/task-list/content.test.ts
@@ -105,9 +105,11 @@ describe('testcase for tasklist', () => {
         tasks: [
           {
             disabled: false,
+            hintText: null,
             href: undefined,
             id: 'childArrangementApplication',
             linkText: 'Your child arrangements application',
+            openInAnotherTab: undefined,
             stateTag: { className: 'govuk-tag--yellow', label: 'In progress' },
           },
         ],
@@ -222,9 +224,11 @@ describe('testcase for tasklist', () => {
         tasks: [
           {
             disabled: false,
+            hintText: null,
             href: undefined,
             id: 'childArrangementApplication',
             linkText: 'Eich cais trefniadau plant',
+            openInAnotherTab: undefined,
             stateTag: { className: 'govuk-tag--yellow', label: 'Ar y gweill' },
           },
         ],
